Wire Redis scheduling, extract link discovery
